codex - note nav drawer in changelog
