+++ docs/src/pages/Installation.tsx
@@ -3,6 +3,7 @@
 // Getting started installation page
 // ─────────────────────────────────────────────────────────────
 import { Surface, Stack, Typography, Button, Panel } from '@archway/valet';
+import NavDrawer from '../components/NavDrawer';
 import { useNavigate } from 'react-router-dom';
 
 export default function InstallationPage() {
@@ -10,6 +11,7 @@ export default function InstallationPage() {
 
   return (
     <Surface>
+      <NavDrawer />
       <Stack spacing={1} preset="showcaseStack">
         <Typography variant="h2" bold>Installation</Typography>
         <Typography>Install via npm:</Typography>
